refactored AutoInitShutdown attribute
